backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
google-generativeai>=0.8.0
requests==2.31.0
